better pointer casts

+++ src/tests/timer_test_ae.c
@@ -1,12 +1,13 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <time.h>
+#include <stdint.h>
 
 #include "events/ae/ae.h"
 
 long long print(struct aeEventLoop *loop, void *clientData)
 {
-    printf("%d - Hello, World\n", (int)(long)clientData);
+    printf("%lld - Hello, World\n", (long long int)(intptr_t)clientData);
     return -1;
 }
 
@@ -15,9 +16,9 @@ int main( int argc, const char* argv[] )
     aeTimeEvent events[10];
     aeEventLoop *loop = aeCreateEventLoop(10, NULL, AE_DEFAULT_FLAGS, NULL);
     assert(loop);
-    int i;
+    intptr_t i;
     for (i = 0; i < 10; i ++) {
-        aeCreateTimeEvent(loop, &events[i], print, (void *)(long)i, NULL, NULL);
+        aeCreateTimeEvent(loop, &events[i], print, (void *)i, NULL, NULL);
         aeSetTimeoutTimeEvent(loop, &events[i], i*1000);
     }
     aeMain(loop);
